% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/asym_cond_qda_fit.R
\name{asym_cond_qda_fit}
\alias{asym_cond_qda_fit}
\title{Asymmetric Conditional QDA (fitting the model)}
\usage{
asym_cond_qda_fit(
  train_data,
  id = "SID",
  Y1 = "Label_1",
  Y2 = "Label_2",
  X1 = c("T500_1", "T1K_1", "T2K_1", "T3K_1", "T4K_1", "T6K_1", "T8K_1"),
  X2 = c("T500_2", "T1K_2", "T2K_2", "T3K_2", "T4K_2", "T6K_2", "T8K_2"),
  method = c("pooled", "separate"),
  k = 4
)
}
\arguments{
\item{train_data}{A data.frame for the training dataset}

\item{id}{A character for the column name which stores id.}

\item{Y1}{A character shows the column name of label 1.}

\item{Y2}{A character shows the column name of label 2.}

\item{X1}{A string of characters for the column names of predictor set 1.}

\item{X2}{A string of characters for the column names of predictor set 2.}

\item{method}{The method user set to estimate mean and covariance parameters.}

\item{k}{The number of classifications.}
}
\value{
A list to store model parameters.
}
\description{
Asymmetric Conditional QDA (fitting the model)
}
\examples{
library(dplyr)
library(mvtnorm)
data(HearingLoss_simu)
fit <- asym_cond_qda_fit(HearingLoss_simu,
                  id = "id",
                  method = "pooled")
test_data_X <- HearingLoss_simu[1,] \%>\% dplyr::select(-"Label_1", -"Label_2",-"id")
asym_cond_qda_predict(fit, test_data_X)
}
