% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/asym_marg_qda_predict.R
\name{asym_marg_qda_predict}
\alias{asym_marg_qda_predict}
\title{Asymmetric Marginal QDA (predicting new phenotypes)}
\usage{
asym_marg_qda_predict(
  qda_model = NA,
  test_data_X = NA,
  X1 = c("T500_1", "T1K_1", "T2K_1", "T3K_1", "T4K_1", "T6K_1", "T8K_1"),
  X2 = c("T500_2", "T1K_2", "T2K_2", "T3K_2", "T4K_2", "T6K_2", "T8K_2")
)
}
\arguments{
\item{qda_model}{A list output from sym_joint_qda_fit}

\item{test_data_X}{A vector of predictors from the test data}

\item{X1}{A string of characters for the column names of predictor set 1.}

\item{X2}{A string of characters for the column names of predictor set 2.}
}
\value{
Predicted individual level phenotype
}
\description{
Asymmetric Marginal QDA (predicting new phenotypes)
}
\examples{
library(dplyr)
library(mvtnorm)
data(HearingLoss_simu)
fit <- asym_marg_qda_fit(HearingLoss_simu,
                  id = "id",
                  method = "pooled")
test_data_X <- HearingLoss_simu[1,] \%>\% dplyr::select(-"Label_1", -"Label_2",-"id")
asym_marg_qda_predict(fit, test_data_X)
}
