% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simplified_bayes_qda_fit.R
\name{simplified_bayes_qda_fit_Shuffling}
\alias{simplified_bayes_qda_fit_Shuffling}
\title{Fitting Simplified Bayesian QDA model (Shuffling version)}
\usage{
simplified_bayes_qda_fit_Shuffling(
  train_data,
  id = "SID",
  Y1 = "Label_1",
  Y2 = "Label_2",
  X1 = c("T500_1", "T1K_1", "T2K_1", "T3K_1", "T4K_1", "T6K_1", "T8K_1"),
  X2 = c("T500_2", "T1K_2", "T2K_2", "T3K_2", "T4K_2", "T6K_2", "T8K_2"),
  k = 4,
  S = 10
)
}
\arguments{
\item{train_data}{A data.frame for the training dataset}

\item{id}{A character for the column name which stores id.}

\item{Y1}{A character shows the column name of label 1.}

\item{Y2}{A character shows the column name of label 2.}

\item{X1}{A string of characters for the column names of predictor set 1.}

\item{X2}{A string of characters for the column names of predictor set 2.}

\item{k}{The number of classifications.}

\item{S}{The number of shuffles.}
}
\value{
A list to store model parameters.
}
\description{
Fitting Simplified Bayesian QDA model (Shuffling version)
}
