% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simplified_joint_qda_fit.R
\name{simplified_joint_qda_fit}
\alias{simplified_joint_qda_fit}
\title{Fitting Simplified Joint QDA model}
\usage{
simplified_joint_qda_fit(
  train_data,
  id = "SID",
  Y1 = "Label_1",
  Y2 = "Label_2",
  X1 = c("T500_1", "T1K_1", "T2K_1", "T3K_1", "T4K_1", "T6K_1", "T8K_1"),
  X2 = c("T500_2", "T1K_2", "T2K_2", "T3K_2", "T4K_2", "T6K_2", "T8K_2"),
  k = 4,
  Shuffle = TRUE,
  S = 10,
  method = c("pooled", "separate")
)
}
\arguments{
\item{train_data}{A data.frame for the training dataset}

\item{id}{A character for the column name which stores id.}

\item{Y1}{A character shows the column name of label 1.}

\item{Y2}{A character shows the column name of label 2.}

\item{X1}{A string of characters for the column names of predictor set 1.}

\item{X2}{A string of characters for the column names of predictor set 2.}

\item{k}{The number of classifications.}

\item{Shuffle}{Whether to use data shuffling before the QDA methods.}

\item{S}{The number of shuffles. Only applicable when using Shuffling.}

\item{method}{The estimation method. Only applicable when not using Shuffling.}
}
\value{
A list to store model parameters.
}
\description{
Fitting Simplified Joint QDA model
}
\examples{
library(dplyr)
library(mvtnorm)
data(HearingLoss_simu)
#Use the Shuffling method
fit <- simplified_joint_qda_fit(HearingLoss_simu,
                  id = "id",
                  Shuffle = TRUE)

#Use the Non-shuffling method
fit <- simplified_joint_qda_fit(HearingLoss_simu,
                  id = "id",
                  Shuffle = FALSE,
                  method = "pooled")
}
