% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sym_joint_qda_fit.R
\name{sym_joint_qda_fit}
\alias{sym_joint_qda_fit}
\title{Symmetric Joint QDA (fitting model)}
\usage{
sym_joint_qda_fit(
  train_data,
  id = "SID",
  Ear1_mark = "_1",
  Ear2_mark = "_2",
  Y = "Label",
  number_features = 7
)
}
\arguments{
\item{train_data}{A data.frame for the training dataset}

\item{id}{A character for the column name which stores id.}

\item{Ear1_mark}{A character shows the last several letters to denote ear 1 label.}

\item{Ear2_mark}{A character shows the last several letters to denote ear 2 label.}

\item{Y}{A character denotes the first several letters of label.}

\item{number_features}{Number of features (covariates) to fit the model for each ear.}
}
\value{
A list to store model parameters.
}
\description{
Symmetric Joint QDA (fitting model)
}
\examples{
library(dplyr)
data(HearingLoss_simu)
fit <- sym_joint_qda_fit(HearingLoss_simu,
                  id = "id")
}
