% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sym_joint_qda_predict.R
\name{sym_joint_qda_predict}
\alias{sym_joint_qda_predict}
\title{Symmetric Joint QDA (predicting new phenotypes)}
\usage{
sym_joint_qda_predict(
  qda_model = NA,
  test_data_X = NA,
  Ear1_mark = "_1",
  Ear2_mark = "_2",
  number_features = 7
)
}
\arguments{
\item{qda_model}{A list output from sym_joint_qda_fit}

\item{test_data_X}{A vector of predictors from the test data}

\item{Ear1_mark}{A character shows the last several letters to denote ear 1 label.}

\item{Ear2_mark}{A character shows the last several letters to denote ear 2 label.}

\item{number_features}{Number of features (covariates) to fit the model for each ear.}
}
\value{
Predicted individual level phenotype
}
\description{
Symmetric Joint QDA (predicting new phenotypes)
}
\examples{
library(dplyr)
library(mvtnorm)
data(HearingLoss_simu)
fit <- sym_joint_qda_fit(HearingLoss_simu,
                  id = "id")
test_data_X <- HearingLoss_simu[1,] \%>\% dplyr::select(-"Label_1", -"Label_2",-"id")
sym_joint_qda_predict(fit, test_data_X)
}
