% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/asym_cond_qda_predict.R
\name{asym_cond_qda_predict}
\alias{asym_cond_qda_predict}
\title{Asymmetric Conditional QDA (predicting new phenotypes)}
\usage{
asym_cond_qda_predict(
  qda_model = NA,
  test_data_X = NA,
  Ear1_mark = "_1",
  Ear2_mark = "_2",
  number_features = 7,
  iter = 1000
)
}
\arguments{
\item{qda_model}{A list output from asym_cond_qda_fit}

\item{test_data_X}{A vector of predictors from the test data}

\item{Ear1_mark}{A character shows the last several letters to denote ear 1 label.}

\item{Ear2_mark}{A character shows the last several letters to denote ear 2 label.}

\item{number_features}{Number of features (covariates) to fit the model for each ear.}

\item{iter}{Number of interations set by users. The default is 1000 times}
}
\value{
Predicted individual level phenotype
}
\description{
Asymmetric Conditional QDA (predicting new phenotypes)
}
\examples{
library(dplyr)
library(mvtnorm)
data(HearingLoss_simu)
fit <- asym_cond_qda_fit(HearingLoss_simu,
                  id = "id")
test_data_X <- HearingLoss_simu[1,] \%>\% dplyr::select(-"Label_1", -"Label_2",-"id")
asym_cond_qda_predict(fit, test_data_X)
}
